% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{generate_absql_runner}
\alias{generate_absql_runner}
\title{Generate an ABSQL Runner}
\usage{
generate_absql_runner(
  extra_constructors = NULL,
  replace_only = FALSE,
  file_context_from = NULL,
  partial_kwargs = NULL,
  ...
)
}
\arguments{
\item{extra_constructors}{Name list of functions that can be used as YAML constructors.}

\item{replace_only}{TRUE if only text replace should be used when rendering text or files.}

\item{file_context_from}{Set to a key name in file frontmatter that should be "promoted" to top-level access}

\item{partial_kwargs}{A vector a keyword arguments that should be partialized in functions when rendering text or files.}

\item{...}{Named arguments for additional context which can be used when rendering text or files.}
}
\value{
ABSQL Runner object
}
\description{
Generate an ABSQL Runner
}
